fix - changing the type of dropout value from int to double

+++ Project/include/model/Model.h
@@ -9,7 +9,7 @@ namespace model {
         FaceRecognitionModelImpl(
             int num_channel,
             int out_dim,
-            int dropout
+            double dropout
         );
 
         torch::Tensor forward(torch::Tensor x);
