+++ Project/include/model/projector.h
@@ -7,7 +7,7 @@ namespace model {
         FaceRecognitionProjectorImpl(
             int in_channel, 
             int out_dim,
-            int dropout
+            double dropout
         );
 
         torch::Tensor forward(torch::Tensor x);
